admin: refactor dashboard page
